e59_agent_mesh/requirements.txt
# Core async HTTP client
httpx==0.24.1

# Web framework
fastapi==0.104.1
uvicorn==0.24.0
aiohttp==3.9.1

# Data handling
numpy==1.24.3
pandas==2.0.3

# Machine Learning
scikit-learn==1.3.1
xgboost==2.0.2

# Scientific computing
scipy==1.11.3

# Database
supabase==2.0.3

# Notifications
twilio==8.10.0
python-dotenv==1.0.0

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
mypy==1.7.1
flake8==6.1.0

# System integration
systemd-python==235  # For sd_notify

# JSON handling (built-in, but explicit for clarity)
# json - stdlib

# Logging (built-in)
# logging - stdlib
